+++ src/components/editor/codemirror/use-codemirror.ts
@@ -21,13 +21,8 @@ import {
 import { useEditorConfig } from '../universal/use-editor-setting'
 import { codemirrorReconfigureExtension } from './extension'
 import { syntaxTheme } from './syntax-highlight'
-import { useCodeMirrorAutoToggleTheme } from './ui'
-
-export const customTheme = EditorView.theme({
-  '&': {
-    height: '100%',
-  },
-})
+import { useCodeMirrorConfigureFonts } from './use-auto-fonts'
+import { useCodeMirrorAutoToggleTheme } from './use-auto-theme'
 
 interface Props {
   initialDoc: string
@@ -147,6 +142,7 @@ export const useCodeMirror = <T extends Element>(
   })
 
   useCodeMirrorAutoToggleTheme(editorView)
+  useCodeMirrorConfigureFonts(editorView)
 
   onBeforeUnmount(() => {
     editorView.value?.destroy()