+++ src/utils/rest.ts
@@ -1,9 +1,9 @@
 import camelcaseKeys from 'camelcase-keys'
 import { API_URL } from 'constants/env'
 import { isPlainObject } from 'lodash-es'
-import type { RequestMethod, RequestOptionsInit } from 'umi-request'
 import { extend } from 'umi-request'
 import { uuid } from 'utils'
+import type { RequestMethod, RequestOptionsInit } from 'umi-request'
 
 import { router } from '../router/router'
 import { getToken } from './auth'