+++ src/views/manage-pages/write.tsx
@@ -19,6 +19,8 @@ import { useRoute, useRouter } from 'vue-router'
 
 import { Icon } from '@vicons/utils'
 
+import { HeaderPreviewButton } from '~/components/special-button/preview'
+
 type PageReactiveType = WriteBaseType & {
   subtitle: string
   slug: string
@@ -104,6 +106,7 @@ const PageWriteView = defineComponent(() => {
 
   return () => (
     <ContentLayout
+      headerClass="pt-1"
       actionsElement={
         <>
           <ParseContentButton
@@ -117,6 +120,8 @@ const PageWriteView = defineComponent(() => {
               data.meta = { ...rest }
             }}
           />
+
+          <HeaderPreviewButton getData={() => ({ ...data })} />
 
           <HeaderActionButton
             icon={<TelegramPlaneIcon />}
@@ -147,7 +152,7 @@ const PageWriteView = defineComponent(() => {
         }}
       ></MaterialInput>
 
-      <div class={'text-gray-700 pt-3'}>
+      <div class={'text-gray-700 dark:text-gray-300 pt-3'}>
         <UnderlineInput
           value={data.subtitle}
           onChange={(e) => void (data.subtitle = e)}